feat: useInputCheck 훅으로 로직 분리
